refactor: modularize App sections
